requirements.txt
asgiref==3.3.4
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-installed-apps-command==2020.12.3
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.1.0
importlib-metadata==3.10
Markdown==3.3.4
psycopg2==2.8.6
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
whitenoise==5.2.0
zipp==3.4.1
